[vim] refactor snippets.lua
