+++ vim/.config/nvim/lua/plugins/snippets.lua
@@ -1,0 +1,81 @@
+-- REFERENCES
+--
+--     https://github.com/L3MON4D3/LuaSnip
+--     https://github.com/honza/vim-snippets/blob/master/snippets/
+--
+
+return {
+    {
+        "L3MON4D3/LuaSnip",
+        config = function()
+            require("luasnip.loaders.from_snipmate").lazy_load()
+
+            local ls = require("luasnip")
+            local s = ls.snippet
+            local t = ls.text_node
+            local i = ls.insert_node
+            local f = ls.function_node
+
+            local function box(opts)
+                -- NOTE: this had to be a function, otherwise `textwidth` was 0 during
+                -- initialization. Still need to better understand when variables are
+                -- populated.
+                local function box_width()
+                    return opts.box_width or vim.opt.textwidth:get()
+                end
+
+                local function padding(cs, input_text)
+                    local spaces = box_width() - (2 * #cs)
+                    spaces = spaces - #input_text
+                    return spaces / 2
+                end
+
+                -- https://github.com/L3MON4D3/LuaSnip/issues/151#issuecomment-912641351
+                local comment_string = function()
+                    return require("luasnip.util.util").buffer_comment_chars()[1]
+                end
+
+                return {
+                    f(function()
+                        local cs = comment_string()
+                        return string.rep(string.sub(cs, 1, 1), box_width())
+                    end, { 1 }),
+                    t({ "", "" }),
+                    f(function(args)
+                        local cs = comment_string()
+                        return cs .. string.rep(" ", math.floor(padding(cs, args[1][1])))
+                    end, { 1 }),
+                    i(1, "placeholder"),
+                    f(function(args)
+                        local cs = comment_string()
+                        return string.rep(" ", math.ceil(padding(cs, args[1][1]))) .. cs
+                    end, { 1 }),
+                    t({ "", "" }),
+                    f(function()
+                        local cs = comment_string()
+                        return string.rep(string.sub(cs, 1, 1), box_width())
+                    end, { 1 }),
+                }
+            end
+
+            require("luasnip").add_snippets("all", {
+                s({ trig = "box" }, box({ box_width = 24 })),
+                s({ trig = "bbox" }, box({})),
+            })
+
+            vim.cmd([[
+imap <silent><expr> <Tab> luasnip#expand_or_jumpable() ? '<Plug>luasnip-expand-or-jump' : '<Tab>'
+inoremap <silent> <S-Tab> <cmd>lua require'luasnip'.jump(-1)<Cr>
+
+snoremap <silent> <Tab> <cmd>lua require('luasnip').jump(1)<Cr>
+snoremap <silent> <S-Tab> <cmd>lua require('luasnip').jump(-1)<Cr>
+]])
+
+            vim.api.nvim_create_user_command(
+                "EditSnippets",
+                'lua require("luasnip.loaders").edit_snippet_files()',
+                { desc = "Edit LuaSnip snippets" }
+            )
+        end,
+    },
+}
